+++ tests/Method/HttpCookieTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Auth\Tests\Method;
 
-use Nyholm\Psr7\Response;
 use Nyholm\Psr7\ServerRequest;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
@@ -58,15 +57,6 @@ final class HttpCookieTest extends TestCase
                 $this->createRequest(['access-token' => 'access-token-value']),
             ),
         );
-    }
-
-    public function testChallengeImmutabilityStatus(): void
-    {
-        $response = new Response(400);
-        $identityRepository = new FakeIdentityRepository($this->createIdentity());
-        $authenticationMethod = new HttpCookie($identityRepository);
-
-        $this->assertSame($response, $authenticationMethod->challenge($response));
     }
 
     public function testCustomTokenParam(): void